Drop sensors map to specific device level, as it static final, it will be shared by all device instances on the gateway, and this was was causing some issue.
